+++ lib/widgets/appbar.dart
@@ -1,18 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
-enum AppBarType {
-  BACK, MAIN
-}
-
-AppBar bircaAppBar(
-    void Function() onPressed
-) => AppBar(
+AppBar bircaAppBar(void Function() onPressed) => AppBar(
     scrolledUnderElevation: 0,
     leading: IconButton(
         onPressed: onPressed,
-        icon: SvgPicture.asset(
-            'lib/assets/image/ic_back.svg'
-        )
-      )
-  );
+        icon: SvgPicture.asset('lib/assets/image/ic_back.svg')));
